Add widgetMetadata that contains mapping of title and group
