+++ src/redux/initial-state/dev-lang.json
@@ -826,7 +826,7 @@
         "-77657645": "Failed to fetch lead groups.",
         "-80733191": "The order of the widgets below represents the order in which the conditions are checked.",
         "-17190550": "You can add conditions from these widgets that are currently in your framework.",
-        "-50186668": "Add from geo",
+        "-50186668": "Geo",
         "-15095273": "Select Gallery Files",
         "-75331863": "Fatal error occured while getting users gallery files",
         "-41121825": "Failed to get user gallery files",
@@ -900,7 +900,13 @@
         "-17378901": "Slack",
         "-49506348": "User Guide",
         "-85406765": "See DEEP User Guide",
-        "-56453375": "User Guide"
+        "-56453375": "User Guide",
+        "-39845084": "Add rows from",
+        "-50010134": "Add cells from",
+        "-96186123": "Add dimensions from",
+        "-63120597": "Add sectors from",
+        "-89164469": "Add sub dimensions from",
+        "-11158375": "Add sub sectors from"
     },
     "links": {
         "browserExtension": {
@@ -2306,21 +2312,18 @@
             "rowTitle": 748,
             "duplicationError": 1008,
             "atLeastOneError": 1007,
-            "addRowButtonTitle": 749,
+            "addRowButtonTitle": 20,
             "cancelButtonLabel": 61,
             "deleteButtonTooltip": -93592403,
             "unnamedRowTitle": -47041993,
-            "addCellButtonTitle": 720,
+            "addCellButtonTitle": 20,
             "cellsHeaderTitle": -1855073,
             "tooltipLabel": 733,
             "unnamedCellTitle": -17919979,
             "removeCellButtonTitle": -53055304,
-            "addFromWidgets": -14972110,
             "cancelConfirmMessage": -14675920,
-            "duplicatesConfirmTitle": -47600296,
-            "duplicatesConfirmText": -76890842,
-            "duplicateConfirmText": -35530735,
-            "removeCellButtonTitle": -53055304
+            "addRowsFromTitle": -39845084,
+            "addCellsFromTitle": -50010134
         },
         "widgets.editor.matrix2d": {
             "atLeastOneError": 1007,
@@ -2331,12 +2334,12 @@
             "saveButtonLabel": 25,
             "sectorsHeaderTitle": 9026,
             "dimensionsHeaderTitle": -25010734,
-            "addSectorButtonTitle": -43570294,
-            "addDimensionButtonTitle": 789,
+            "addSectorButtonTitle": 20,
+            "addDimensionButtonTitle": 20,
             "colorLabel": 764,
             "tooltipLabel": 733,
             "unnamedDimensionLabel": -52234257,
-            "addSubdimensionButtonTitle": 788,
+            "addSubdimensionButtonTitle": 20,
             "subdimensionsHeaderTitle": 786,
             "unnamedSubdimensionLabel": -7983483,
             "removeSubdimensionButtonTooltip": -76009053,
@@ -2344,15 +2347,15 @@
             "unnamedSectorTitle": -8790199,
             "deleteSectorTooltip": -64784286,
             "unnamedSectorLabel": -8790199,
-            "addSubsectorButtonTitle": -59828541,
+            "addSubsectorButtonTitle": 20,
             "subsectorsHeaderTitle": 7039,
             "unnamedSubsectorLabel": -13931386,
             "removeSubsectorButtonTooltip": -24172161,
-            "addFromWidgets": -14972110,
             "cancelConfirmMessage": -14675920,
-            "duplicateConfirmText": -35530735,
-            "duplicatesConfirmText": -76890842,
-            "removeSubsectorButtonTooltip": -24172161
+            "addDimensionsTitle": -96186123,
+            "addSectorsTitle": -63120597,
+            "addSubDimensionsTitle": -89164469,
+            "addSubSectorTitle": -11158375
         },
         "editEntry.list": {
             "noEntriesText": 543
@@ -2390,8 +2393,7 @@
             "rootNodeLabel": 5552,
             "addFromGeoLabel": -50186668,
             "cancelConfirmMessage": -14675920,
-            "addFromWidgets": -14972110,
-            "duplicateConfirmText": -35530735,
+            "addFromWidgets": -66364403,
             "duplicatesConfirmText": -76890842,
             "duplicatesConfirmTitle": -47600296,
             "nonDuplicatesConfirmText": -867218
