+++ src/widgets/edit/Conditional/styles.scss
@@ -1,7 +1,7 @@
 @import '~base-scss/base';
 
 $height-modal-body: 720px;
-$width-modal: 720px;
+$width-modal: 960px;
 
 .sortable-unit {
     border-top: $width-separator-thin solid $color-separator;
