+++ src/widgets/edit/Matrix1d/RowContent/styles.scss
@@ -27,6 +27,14 @@
     .button-container {
         display: flex;
         align-items: center;
+
+        h5 {
+            padding-right: $spacing-small;
+
+            &:after {
+                content: ':';
+            }
+        }
     }
 }
