+++ src/widgets/edit/Matrix2d/SectorContent/styles.scss
@@ -27,6 +27,14 @@
     .button-container {
         display: flex;
         align-items: center;
+
+        h5 {
+            padding-right: $spacing-small;
+
+            &:after {
+                content: ':';
+            }
+        }
     }
 }
