+++ src/widgets/index.js
@@ -4,6 +4,7 @@ import update from '#rsu/immutable-update';
 
 import FrameworkWidgetError from '#components/FrameworkWidgetError';
 import WidgetError from '#components/WidgetError';
+import { widgetTitlesGroupMap } from '#widgets/widgetMetadata';
 
 import DefaultEditWidget from './edit/Default';
 // overview
@@ -75,47 +76,8 @@ export const VIEW = {
 const editDecorator = boundError(FrameworkWidgetError);
 const decorator = boundError(WidgetError);
 
-export const widgetGroups = {
-    // NOTE: misc is a special kind of widgetGroup (should always be defined)
-    misc: {
-        // NOTE: used as _ts('widgetGroupTitle', 'miscGroupLabel')
-        title: 'miscGroupLabel',
-        order: 5,
-    },
-
-    excerpts: {
-        // NOTE: used as _ts('widgetGroupTitle', 'excerptsGroupLabel')
-        title: 'excerptsGroupLabel',
-        order: 1,
-    },
-    matrices: {
-        // NOTE: used as _ts('widgetGroupTitle', 'matricesGroupLabel')
-        title: 'matricesGroupLabel',
-        order: 2,
-    },
-    temporals: {
-        // NOTE: used as _ts('widgetGroupTitle', 'temporalsGroupLabel')
-        title: 'temporalsGroupLabel',
-        order: 3,
-    },
-    selections: {
-        // NOTE: used as _ts('widgetGroupTitle', 'selectionsGroupLabel')
-        title: 'selectionsGroupLabel',
-        order: 4,
-    },
-    conditionals: {
-        // NOTE: used as _ts('widgetGroupTitle', 'conditionalsGroupLabel')
-        title: 'conditionalsGroupLabel',
-        order: 6,
-    },
-};
-
 export const globalWidgets = {
     excerptWidget: {
-        groupId: 'excerpts',
-
-        // NOTE: used as _ts('widgetTitle', 'excerptWidgetLabel')
-        title: 'excerptWidgetLabel',
         editComponent: DefaultEditWidget,
 
         overview: {
@@ -130,10 +92,6 @@ export const globalWidgets = {
     },
 
     matrix1dWidget: {
-        groupId: 'matrices',
-
-        // NOTE: used as _ts('widgetTitle', 'matrix1DWidgetLabel')
-        title: 'matrix1DWidgetLabel',
         editComponent: Matrix1dEditWidget,
 
         overview: {
@@ -147,10 +105,6 @@ export const globalWidgets = {
         },
     },
     matrix2dWidget: {
-        groupId: 'matrices',
-
-        // NOTE: used as _ts('widgetTitle', 'matrix2DWidgetLabel')
-        title: 'matrix2DWidgetLabel',
         editComponent: Matrix2dEditWidget,
 
         overview: {
@@ -164,10 +118,6 @@ export const globalWidgets = {
         },
     },
     numberMatrixWidget: {
-        groupId: 'matrices',
-
-        // NOTE: used as _ts('widgetTitle', 'numberMatrixWidgetLabel')
-        title: 'numberMatrixWidgetLabel',
         editComponent: NumberMatrixEditWidget,
 
         overview: {
@@ -183,10 +133,6 @@ export const globalWidgets = {
 
 
     dateWidget: {
-        groupId: 'temporals',
-
-        // NOTE: used as _ts('widgetTitle', 'dateWidgetLabel')
-        title: 'dateWidgetLabel',
         editComponent: DefaultEditWidget,
 
         overview: {
@@ -201,10 +147,6 @@ export const globalWidgets = {
         },
     },
     timeWidget: {
-        groupId: 'temporals',
-
-        // NOTE: used as _ts('widgetTitle', 'timeWidgetLabel')
-        title: 'timeWidgetLabel',
         editComponent: DefaultEditWidget,
 
         overview: {
@@ -219,10 +161,6 @@ export const globalWidgets = {
         },
     },
     timeRangeWidget: {
-        groupId: 'temporals',
-
-        // NOTE: used as _ts('widgetTitle', 'timeRangeWidgetLabel')
-        title: 'timeRangeWidgetLabel',
         editComponent: DefaultEditWidget,
 
         overview: {
@@ -237,10 +175,6 @@ export const globalWidgets = {
         },
     },
     dateRangeWidget: {
-        groupId: 'temporals',
-
-        // NOTE: used as _ts('widgetTitle', 'dateRangeWidgetLabel')
-        title: 'dateRangeWidgetLabel',
         editComponent: DefaultEditWidget,
 
         overview: {
@@ -256,10 +190,6 @@ export const globalWidgets = {
     },
 
     numberWidget: {
-        groupId: 'misc',
-
-        // NOTE: used as _ts('widgetTitle', 'numberWidgetLabel')
-        title: 'numberWidgetLabel',
         editComponent: NumberEditWidget,
 
         overview: {
@@ -274,10 +204,6 @@ export const globalWidgets = {
         },
     },
     scaleWidget: {
-        groupId: 'misc',
-
-        // NOTE: used as _ts('widgetTitle', 'scaleWidgetLabel')
-        title: 'scaleWidgetLabel',
         editComponent: ScaleEditWidget,
 
         overview: {
@@ -292,10 +218,6 @@ export const globalWidgets = {
         },
     },
     geoWidget: {
-        groupId: 'misc',
-
-        // NOTE: used as _ts('widgetTitle', 'geoWidgetLabel')
-        title: 'geoWidgetLabel',
         editComponent: DefaultEditWidget,
 
         overview: {
@@ -310,10 +232,6 @@ export const globalWidgets = {
         },
     },
     organigramWidget: {
-        groupId: 'misc',
-
-        // NOTE: used as _ts('widgetTitle', 'organigramWidgetLabel')
-        title: 'organigramWidgetLabel',
         editComponent: OrganigramEditWidget,
 
         overview: {
@@ -329,10 +247,6 @@ export const globalWidgets = {
     },
 
     selectWidget: {
-        groupId: 'selections',
-
-        // NOTE: used as _ts('widgetTitle', 'selectWidgetLabel')
-        title: 'selectWidgetLabel',
         editComponent: MultiSelectEditWidget,
 
         overview: {
@@ -347,10 +261,6 @@ export const globalWidgets = {
         },
     },
     multiselectWidget: {
-        groupId: 'selections',
-
-        // NOTE: used as _ts('widgetTitle', 'multiselectWidgetLabel')
-        title: 'multiselectWidgetLabel',
         editComponent: MultiSelectEditWidget,
 
         overview: {
@@ -366,10 +276,6 @@ export const globalWidgets = {
     },
 
     conditionalWidget: {
-        groupId: 'conditionals',
-
-        // NOTE: used as _ts('widgetTitle', 'conditionalWidgetLabel')
-        title: 'conditionalWidgetLabel',
         editComponent: ConditionalEditWidget,
         frameworkComponent: ConditionalFrameworkPreview,
 
@@ -390,28 +296,22 @@ export const globalWidgets = {
 const widgets = mapToMap(
     globalWidgets,
     undefined,
-    (widget) => {
+    (widget, widgetKey) => {
         const {
-            title,
             editComponent,
             frameworkComponent,
             list,
             overview,
-            groupId = 'misc',
         } = widget;
+        const {
+            title,
+            groupId = 'misc',
+        } = widgetTitlesGroupMap[widgetKey];
 
         const prepareMinSize = ms => ({
             width: gridSize.width * ms.w,
             height: gridSize.height * ms.h,
         });
-
-        /*
-        let groupInfo = widgetGroups[groupId];
-        if (!groupInfo) {
-            console.error(`Undefined group ${groupId} for widget`);
-            groupInfo = widgetGroups.misc;
-        }
-        */
 
         const componentSettings = {
             initialLayout: { $set: {
